Introduce LabelConfigurator, remove label extension
