--- Sources/GravatarUI/DesignSystem/PaletteRefreshable.swift
@@ -1,5 +1,0 @@
-
-/// Defines a type that can adapt to ``PaletteType`` changes.
-public protocol PaletteRefreshable {
-    func refresh(with paletteType: PaletteType)
-}
