+++ Sources/GravatarUI/GravatarCompatibleUI/GravatarCompatible.swift
@@ -21,4 +21,3 @@ extension GravatarCompatible {
 }
 
 extension UIImageView: GravatarCompatible {}
-extension UILabel: GravatarCompatible {}
